fastapi_firestore/requirements.txt
fastapi==0.115.11
joblib
numpy==2.2.3
protobuf==6.30.1
pydantic==2.10.6
uvicorn
google-cloud-firestore
scikit-learn